fixes and changes
removed blocking behavior in sending packets: now the packets no longer check if the partition of destination is allocated on the receiving node (this created implicit deadlocks for particular configurations). Now, the messages are sent as soon as their dependency gets cleared (like before). On the other hand, more logic on the memory side was needed